Cleanup. With Lint. Redundant images.

+++ TradeHero/TH_ANDROID/src/googleplay/java/com/tradehero/common/billing/googleplay/IABConstants.java
@@ -140,7 +140,7 @@ public class IABConstants implements BillingConstants
         }
         else if (o instanceof Integer)
         {
-            return ((Integer) o).intValue();
+            return (Integer) o;
         }
         else if (o instanceof Long)
         {
